FOGL-904 made common microservice_management_client class
